refactor(parser_gateway): split main.rs into modules, wire x402 gating + attestation + bearer auth
main.rs had grown into one file holding the Turnkey envelope
construction, the health and parse handlers, and all of main()'s
wiring. Adding x402 gating, TVC attestation, and the bearer-token gate
on top of that would have made it worse. Split into state.rs
(AppState), handlers/{health,parse}.rs, plus the attestation/auth/
x402_config modules landed in the previous three commits, so main()
is now just: build the attestation verifier, build the router, probe
the x402 facilitator and mount /v2 behind its middleware if that
succeeds, layer the optional bearer-token gate, serve.

The mock boot-proof handling (MOCK_BOOT_PROOF_* constants,
mock_boot_proof(), and the error_response(msg) shim over
host_primitives::turnkey::error_response) moves into handlers/parse.rs,
since that's the only remaining consumer; the wire-shape tests
(mock_boot_proof_matches_production_wire_shape,
error_response_carries_mock_boot_proof) move with it unchanged, same
mock constant values, same assertions.

Bumps axum 0.6.20 -> 0.8 for this crate (parser/cli already runs 0.8,
so two axum majors in the workspace is the existing status quo) and
migrates main() off the removed axum::Server API onto
tokio::net::TcpListener + axum::serve.

v1 stays open; v2 is gated by x402-axum's middleware built from
x402_config, mounted only if the configured facilitator answers a
/supported probe at startup. Route logic, timeouts, and error handling
are otherwise unchanged from what main.rs already did.

Ported from the unmerged spec/x402-tvc-enforced branch (PR #304) at
its pre-VPM state (commits 8124e712, 7c144656, f323def7), since the
later TVC-enforced handler needs the gateway-signed payment-marker
types that a parallel PR is still landing. Also carries bd3b0657's
64 KiB public body cap.
